% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CheckPoint}
\alias{CheckPoint}
\title{Check if points in segments}
\usage{
CheckPoint(pos, start, end)
}
\arguments{
\item{pos}{a vector of numbers of points}

\item{start}{a vector of numbers of starting position}

\item{end}{a vector of numbers of ending position}
}
\description{
To check if a number in a range. 
Return a vector of 0,1 if the the number 
in a range(1) or not (0).
}
\examples{
aa <- c(2,4,6)
bb1 <- c(1,5,7)
bb2 <- c(3,8,10)
#too see if 2,4,6 in range 1-3,5-8,7-10
CheckPoint(aa,bb1,bb2)
}
