% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CheckRange}
\alias{CheckRange}
\title{Check if segments contain point}
\usage{
CheckRange(pos, start, end)
}
\arguments{
\item{pos}{a vector of numbers of points}

\item{start}{a vector of numbers of starting position}

\item{end}{a vector of numbers of ending position}
}
\description{
To check if a range contains points. 
Return a vector of 0,1 
with 1 = contain points, 0 = not contain points.
}
\examples{
aa <- c(2,4,6)
bb1 <- c(1,5,7)
bb2 <- c(3,8,10)
# to see if range 1-3,5-8, 7-8 contain 2,4,6
CheckRange(aa,bb1,bb2)
}
