% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CheckRange2Range}
\alias{CheckRange2Range}
\title{Check if segments overlap with other segments}
\usage{
CheckRange2Range(check_start, check_end, other_start, other_end)
}
\arguments{
\item{check_start}{a vector of numbers of starting positions of checked segments}

\item{check_end}{a vector of numbers of ending position of checked segments}

\item{other_start}{a vector of numbers of starting position of other segments}

\item{other_end}{a vector of numbers of ending position of other segments}
}
\description{
To check if a segments overlap with other segments.

Return a vector of 0,1
with 1 = overlap, 0 = not overlap.
}
\examples{
aa_start <- c(2,6,11)
aa_end <- c(4,9,15)
bb_start <- c(1,5,7)
bb_end <- c(3,8,10)
# to see if segments 2-4,6-9,11-15 overlap with 1-3,5-8,7-10
CheckRange2Range(aa_start,aa_end,bb_start,bb_end)
}
