% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CheckMatch}
\alias{CheckMatch}
\title{match point to range}
\usage{
CheckMatch(pos, start, end)
}
\arguments{
\item{pos}{a vector of numbers of points}

\item{start}{a vector of numbers of starting position}

\item{end}{a vector of numbers of ending position}
}
\description{
To map points into segmentatin
return a dataframe with "point_position" and "seg_row"
indicating the points fall into which segmentation.
}
\examples{
aa <- c(3,9,21,11)
bb1 <- c(1,8,16)
bb2 <- c(4,15,18)
CheckMatch(aa,bb1,bb2)
#result shows the first point in first segmentation; the second and forth points
#in the second segmentation
}
